feat: Enhance COS trainer with expert-parallel metric exchange and update logging format; add configuration files for COS and SFT training

+++ src/models/wan_i2v.py
@@ -440,7 +440,7 @@ class WanI2VForTraining:
             sigmas = shifted_sigmas.index_select(0, indices)
             timesteps = shifted_timesteps.index_select(0, indices)
             weights = bsmntw.index_select(0, indices)
-            sigmas_5d = sigmas.view(B, 1, 1, 1, 1)
+            sigmas_5d = sigmas.view(B, 1, 1, 1, 1).to(x_final.dtype)
 
             noise = torch.randn_like(x_final)
